Clean up component stories
