+++ src/components/FormatIcon/Icon.jsx
@@ -1,5 +1,3 @@
-/* eslint-disable */
-
 import React from 'react'
 import defaultTheme from '../../theme/default'
