+++ src/components/FormatIcon/index.jsx
@@ -1,5 +1,3 @@
-/* eslint-disable */
-
 import React from 'react'
 import Icon from './Icon'
